+++ clients/fides-js/src/lib/common-utils.ts
@@ -1,0 +1,3 @@
+export const raise = (message: string) => {
+  throw new Error(message);
+};